react-agent/requirements.txt
# LangGraph & LangChain Core
langgraph>=0.2.0
langgraph-cli[inmem]>=0.1.71
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0

# RAG & Vector Store
chromadb>=0.4.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# Document Processing
pypdf>=3.0.0
python-docx>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
httpx>=0.25.0

# Caching
diskcache>=5.6.0

# Environment
python-dotenv>=1.0.0

# Search (optional)
tavily-python>=0.3.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
